refactor(cli): extract resolveTargetAssistant into lib module
cli/CONTRIBUTING.md says shared logic belongs in src/lib/ rather
than being exported from src/commands/. Move the upgrade-specific
resolveTargetAssistant (with platform fallback) from
src/commands/upgrade.ts into a new src/lib/upgrade-resolve.ts so
the function isn't exported from a command file. The test file
import path follows.

Addresses Devin review feedback on feature branch PR #29680.
